App/requirements.txt
dash==1.11.0
mysql-connector
pymysql
matplotlib
pandas
